src/requirements.txt
Django==1.10
Pillow==3.3.1
django-bootstrap3==7.0.1
django-filter==0.14.0
psycopg2==2.6.2
six==1.10.0
wsgiref==0.1.2
requests==2.11.1
django-geojson==2.9.1
django-leaflet==0.20.0
django-autocomplete-light==3.2.0
django-tinymce==2.6.0
svgis==0.4.1
